chore: remove Trip Search v1 components (#2389)
* refactor

* delete!

* refactor

* remove local BFF URL override
