+++ src/screens/Assistant/index.tsx
@@ -2,9 +2,7 @@ import React from 'react';
 import {createStackNavigator, TransitionPresets} from '@react-navigation/stack';
 import {LocationWithMetadata} from '@atb/favorites/types';
 import {DetailsRouteParams} from '../TripDetails/Details';
-import {DetailsRouteParams as DetailsRouteParams_v2} from '../TripDetails_v2/Details';
-import TripDetailsRoot from '../TripDetails';
-import {TripDetailsRoot as TripDetailsRoot_v2} from '../TripDetails_v2';
+import {TripDetailsRoot} from '../TripDetails';
 import AssistantRoot from './Assistant';
 import JourneyDatePicker, {
   DateTimePickerParams,
@@ -18,7 +16,6 @@ export type AssistantParams = {
     searchTime: SearchTime;
   };
   TripDetails: DetailsRouteParams;
-  TripDetails_v2: DetailsRouteParams_v2;
   DateTimePicker: DateTimePickerParams;
 };
 
@@ -37,13 +34,6 @@ const Assistant = () => {
       <Stack.Screen
         name="TripDetails"
         component={TripDetailsRoot}
-        options={{
-          ...TransitionPresets.SlideFromRightIOS,
-        }}
-      ></Stack.Screen>
-      <Stack.Screen
-        name="TripDetails_v2"
-        component={TripDetailsRoot_v2}
         options={{
           ...TransitionPresets.SlideFromRightIOS,
         }}
