+++ src/screens/TripDetails/Details/utils.ts
@@ -1,4 +1,7 @@
 // @TODO Remote config? User settings?
+
+import {Leg} from '@atb/api/types/trips';
+
 const TIME_LIMIT_IN_MINUTES = 3;
 const MIN_SIGNIFICANT_WALK_IN_SECONDS = 30;
 const MIN_SIGNIFICANT_WAIT_IN_SECONDS = 30;
@@ -11,4 +14,12 @@ export function significantWalkTime(seconds: number) {
 }
 export function significantWaitTime(seconds: number) {
   return seconds > MIN_SIGNIFICANT_WAIT_IN_SECONDS;
+}
+
+export function getLineName(leg: Leg) {
+  const name =
+    leg.fromEstimatedCall?.destinationDisplay?.frontText ??
+    leg.line?.name ??
+    '';
+  return leg.line?.publicCode ? `${leg.line.publicCode} ${name}` : name;
 }
