+++ src/screens/TripDetails/Map/types.ts
