--- src/screens/TripDetails_v2/DepartureDetails/types.ts
@@ -1,7 +1,0 @@
-export type ServiceJourneyDeparture = {
-  serviceJourneyId: string;
-  date: string;
-  serviceDate: string;
-  fromQuayId?: string;
-  toQuayId?: string;
-};
